minor documentation edits and renames
