+++ src/components/FileUploadModal/local-upload-button.tsx
@@ -3,7 +3,7 @@ import { Button, Upload, UploadProps } from "antd";
 import { Link } from "react-router-dom";
 import { VIEWER_PATHNAME } from "../../routes";
 
-const LocalUpload: React.FC<UploadProps> = (props) => (
+const LocalUploadButton: React.FC<UploadProps> = (props) => (
     <Upload {...props}>
         <Link
             // Redirect to /viewer if necessary and/or clear out viewer
@@ -17,4 +17,4 @@ const LocalUpload: React.FC<UploadProps> = (props) => (
     </Upload>
 );
 
-export default LocalUpload;
+export default LocalUploadButton;
